# docs/openapi.yaml
openapi: 3.0.3
info:
  title: InvestIQ API
  description: |
    InvestIQ is a comprehensive algorithmic trading platform with multi-engine analysis,
    backtesting, portfolio management, and autonomous trading capabilities.

    ## Authentication
    All endpoints (except `/health` and `/metrics`) require authentication via the `X-API-Key` header or Bearer token.

    ## Live Trading Protection
    Broker write endpoints (POST /api/broker/execute, DELETE /api/broker/positions/:symbol, POST /api/broker/orders/:id/cancel)
    require an additional `X-Live-Trading-Key` header for safety.

    ## Response Format
    All API responses follow a standard wrapper:
    ```json
    {
      "success": true,
      "data": { ... },
      "error": null
    }
    ```
  version: 1.0.0
  contact:
    name: InvestIQ Support
    url: https://github.com/yourusername/invest-iq
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.investiq.io
    description: Production server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/watchlist/personalized:
    get:
      tags: [Watchlist]
      summary: Get personalized opportunity feed
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: min_confidence
          in: query
          schema:
            type: number
            default: 0.3
      responses:
        '200':
          description: Personalized opportunities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/watchlist/interaction:
    post:
      tags: [Watchlist]
      summary: Record a user interaction (for learning)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, symbol, action]
              properties:
                user_id:
                  type: string
                symbol:
                  type: string
                action:
                  type: string
                  enum: [view, click, watchlist_add, watchlist_remove, trade, dismiss]
                context:
                  type: string
      responses:
        '200':
          description: Interaction recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/watchlist/preferences:
    get:
      tags: [Watchlist]
      summary: Get user preferences
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: User preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [Watchlist]
      summary: Update user preferences
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sectors:
                  type: array
                  items:
                    type: string
                risk_tolerance:
                  type: number
                excluded_symbols:
                  type: array
                  items:
                    type: string
                min_confidence:
                  type: number
      responses:
        '200':
          description: Preferences updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/watchlist/items:
    get:
      tags: [Watchlist]
      summary: Get watchlist items
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Watchlist items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [Watchlist]
      summary: Add to watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, symbol]
              properties:
                user_id:
                  type: string
                symbol:
                  type: string
                notes:
                  type: string
                target_price:
                  type: number
                stop_loss:
                  type: number
      responses:
        '200':
          description: Added to watchlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/watchlist/items/{symbol}:
    delete:
      tags: [Watchlist]
      summary: Remove from watchlist
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Removed from watchlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/watchlist/scan:
    get:
      tags: [Watchlist]
      summary: Scan opportunities without personalization
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: min_confidence
          in: query
          schema:
            type: number
            default: 0.3
      responses:
        '200':
          description: List of opportunities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ============================================================================
  # Tax Optimization
  # ============================================================================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication. Can also use Bearer token in Authorization header.

    LiveTradingAuth:
      type: apiKey
      in: header
      name: X-Live-Trading-Key
      description: Additional key required for live trading write operations.

    AdminAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Admin-level API key required for privileged operations.
  schemas:
    ApiResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        data:
          type: object
          description: Response data (structure varies by endpoint)
        error:
          type: string
          nullable: true
...